Add peer handshake for automatic replication
