--- chat/Dockerfile
@@ -1,11 +1,0 @@
-FROM python:3.11-slim
-
-WORKDIR /app
-
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY . .
-
-RUN chmod +x entrypoint.sh
-ENTRYPOINT ["./entrypoint.sh"]
