--- docker-compose.yml
@@ -1,50 +1,0 @@
-version: "3.8"
-services:
-  peer1:
-    build: ./chat
-    container_name: peer1
-    environment:
-      - PORTA_RECEBIMENTO=5001
-      - PEER_NAME=Peer-1
-      - PEER_DESTINO=peer1:5001
-    ports:
-      - "5001:5001"
-    volumes:
-      - logs-volume:/logs
-    networks:
-      - chatmesh-net
-
-  peer2:
-    build: ./chat
-    container_name: peer2
-    environment:
-      - PORTA_RECEBIMENTO=5002
-      - PEER_NAME=Peer-2
-      - PEER_DESTINO=peer2:5002
-    ports:
-      - "5002:5002"
-    volumes:
-      - logs-volume:/logs
-    networks:
-      - chatmesh-net
-
-  peer3:
-    build: ./chat
-    container_name: peer3
-    environment:
-      - PORTA_RECEBIMENTO=5003
-      - PEER_NAME=Peer-3
-      - PEER_DESTINO=peer3:5003
-    ports:
-      - "5003:5003"
-    volumes:
-      - logs-volume:/logs
-    networks:
-      - chatmesh-net
-
-volumes:
-  logs-volume:
-
-networks:
-  chatmesh-net:
-    driver: bridge
